- colored 2.1 → 3.1.1 (transparent)
- notify 7.0 → 8.2.0 (transparent)
- notify-debouncer-full 0.4 → 0.7.0 (transparent)
- tree-sitter 0.24 → 0.26.6 (12 x usize→u32 casts in child() calls)
- tree-sitter-{c,go,lua,php,python,rust,swift} bumped to latest
503 tests passing

+++ src/parser/builtin/python.rs
@@ -341,7 +341,7 @@ impl PythonParser {
             };
 
             for i in 0..body.child_count() {
-                if let Some(child) = body.child(i) {
+                if let Some(child) = body.child(i as u32) {
                     match child.kind() {
                         "function_definition" => {
                             if let Some(entry) =
@@ -353,7 +353,7 @@ impl PythonParser {
                         "decorated_definition" => {
                             // Find the function_definition inside the decorated_definition
                             for j in 0..child.child_count() {
-                                if let Some(inner) = child.child(j) {
+                                if let Some(inner) = child.child(j as u32) {
                                     if inner.kind() == "function_definition" {
                                         if let Some(mut entry) = Self::extract_python_method_entry(
                                             &class_name,
